refactor(search): extract SearchResults.svelte from search page
